Change FarmPlot index to uint8

+++ src/constants.ts
@@ -16,7 +16,7 @@ export const CONTRACT_ABI = parseAbi([
   `struct Campaign { string name; string description; uint256 amount; }`,
   `struct FarmPlot { uint256 time; uint8 plotType; }`,
   `function addCampaign(string memory name, string memory description, uint256 amount) external`,
-  `function setFarmPlot(address user, uint256 index, uint8 plotType) external`,
+  `function setFarmPlot(address user, uint8 index, uint8 plotType) external`,
   `function chargeNextCampaign(uint256 amount) external`,
   `function getNextCampaign() external view returns (Campaign memory)`,
   `function getFarmPlots(address user) external view returns (FarmPlot[] memory)`,
